minor: preview_template_tree迁移到pipeline_web模块
