+++ gcloud/taskflow3/models.py
@@ -23,11 +23,8 @@ from django.db import models, transaction
 from django.utils.translation import ugettext_lazy as _
 
 from pipeline.core.constants import PE
-from pipeline.component_framework.constant import ConstantPool
-from pipeline.models import PipelineInstance, TemplateScheme
+from pipeline.models import PipelineInstance
 from pipeline.engine import states
-from pipeline.validators.gateway import validate_gateways
-from pipeline.validators.utils import format_node_io_to_list
 
 from gcloud.taskflow3.utils import parse_node_timeout_configs
 from pipeline_web.core.abstract import NodeAttr
@@ -562,33 +559,6 @@ class TaskFlowStatisticsMixin(ClassificationCountMixin):
 
 class TaskFlowInstanceManager(models.Manager, TaskFlowStatisticsMixin):
     @staticmethod
-    def sub_process_scheme_pipeline_tree(pipeline_tree):
-        """
-        递归根据执行方案创建子流程的任务实例
-        @return:
-        """
-
-        def _sub_process_scheme_pipeline_tree(pipeline_tree):
-            activities = pipeline_tree[PE.activities]
-
-            for node_id, node in list(activities.items()):
-                if node[PE.type] == PE.SubProcess:
-                    sub_pipeline = node[PE.pipeline]
-                    template_nodes_set = set(sub_pipeline[PE.activities].keys())
-
-                    scheme_id_list = node.get("scheme_id_list", [])
-                    exclude_task_nodes_id = get_template_exclude_task_nodes_with_schemes(
-                        template_nodes_set, scheme_id_list
-                    )
-                    TaskFlowInstanceManager.preview_pipeline_tree_exclude_task_nodes(
-                        sub_pipeline, exclude_task_nodes_id
-                    )
-
-                    _sub_process_scheme_pipeline_tree(sub_pipeline)
-
-        return _sub_process_scheme_pipeline_tree(pipeline_tree)
-
-    @staticmethod
     def create_pipeline_instance(template, **kwargs):
         pipeline_tree = kwargs["pipeline_tree"]
         replace_template_id(template.__class__, pipeline_tree)
@@ -598,8 +568,6 @@ class TaskFlowInstanceManager(models.Manager, TaskFlowStatisticsMixin):
             "description": kwargs.get("description", ""),
         }
         PipelineTemplateWebWrapper.unfold_subprocess(pipeline_tree, template.__class__)
-
-        TaskFlowInstanceManager.sub_process_scheme_pipeline_tree(pipeline_tree)
 
         pipeline_web_cleaner = PipelineWebTreeCleaner(pipeline_tree)
         nodes_attr = pipeline_web_cleaner.clean(with_subprocess=True)
@@ -689,258 +657,6 @@ class TaskFlowInstanceManager(models.Manager, TaskFlowStatisticsMixin):
         pipeline_inst = TaskFlowInstanceManager.create_pipeline_instance(template, **task_info)
 
         return pipeline_inst
-
-    @staticmethod
-    def _replace_node_incoming(next_node, replaced_incoming, new_incoming):
-        if isinstance(next_node[PE.incoming], list):
-            next_node[PE.incoming].pop(next_node[PE.incoming].index(replaced_incoming))
-            next_node[PE.incoming].extend(new_incoming)
-        else:
-            is_boring_list = isinstance(new_incoming, list) and len(new_incoming) == 1
-            next_node[PE.incoming] = new_incoming[0] if is_boring_list else new_incoming
-
-    @staticmethod
-    def _ignore_act(act, locations, lines, pipeline_tree):
-
-        # change next_node's incoming: task node、control node is different
-        # change incoming_flow's target to next node
-        # delete outgoing_flow
-        incoming_id_list, outgoing_id = act[PE.incoming], act[PE.outgoing]
-        incoming_id_list = incoming_id_list if isinstance(incoming_id_list, list) else [incoming_id_list]
-
-        outgoing_flow = pipeline_tree[PE.flows][outgoing_id]
-        target_id = outgoing_flow[PE.target]
-
-        next_node = (
-            pipeline_tree[PE.activities].get(target_id)
-            or pipeline_tree[PE.gateways].get(target_id)
-            or pipeline_tree[PE.end_event]
-        )
-
-        TaskFlowInstanceManager._replace_node_incoming(
-            next_node=next_node, replaced_incoming=outgoing_id, new_incoming=incoming_id_list
-        )
-
-        for incoming_id in incoming_id_list:
-            incoming_flow = pipeline_tree[PE.flows][incoming_id]
-            incoming_flow[PE.target] = next_node["id"]
-
-        pipeline_tree[PE.flows].pop(outgoing_id)
-
-        # web location data
-        try:
-            locations.pop(act["id"])
-            lines.pop(outgoing_id)
-
-            for incoming_id in incoming_id_list:
-                lines[incoming_id][PE.target]["id"] = next_node["id"]
-        except Exception:
-            logger.exception(
-                "create_pipeline_instance_exclude_task_nodes adjust web data error: %s" % traceback.format_exc()
-            )
-
-    @staticmethod
-    def _remove_useless_constants(exclude_task_nodes_id, pipeline_tree):
-        # pop unreferenced constant
-        data = {}
-        for act_id, act in list(pipeline_tree[PE.activities].items()):
-            if act["type"] == PE.ServiceActivity:
-                node_data = {("%s_%s" % (act_id, key)): value for key, value in list(act["component"]["data"].items())}
-            # PE.SubProcess
-            else:
-                node_data = {
-                    ("%s_%s" % (act_id, key)): value
-                    for key, value in list(act.get("constants", {}).items())
-                    if value["show_type"] == "show"
-                }
-
-                sub_pipeline_node_data = {
-                    ("%s_%s" % (act_id, key)): {"hook": True, "value": key}
-                    for key, value in list(pipeline_tree["constants"].items())
-                    if act_id in value["source_info"].keys()
-                }
-                data.update(sub_pipeline_node_data)
-            data.update(node_data)
-
-        for gw_id, gw in list(pipeline_tree[PE.gateways].items()):
-            if gw["type"] in [PE.ExclusiveGateway, PE.ConditionalParallelGateway]:
-                gw_data = {
-                    ("%s_%s" % (gw_id, key)): {"value": value["evaluate"]}
-                    for key, value in list(gw["conditions"].items())
-                }
-                data.update(gw_data)
-
-        # get all referenced constants in flow
-        constants = pipeline_tree[PE.constants]
-
-        referenced_keys = []
-        while True:
-            last_count = len(referenced_keys)
-            cons_pool = ConstantPool(data, lazy=True)
-            refs = cons_pool.get_reference_info(strict=False)
-            for keys in list(refs.values()):
-                for key in keys:
-                    # add outputs keys later
-                    if key in constants and key not in referenced_keys:
-                        referenced_keys.append(key)
-                        data.update({key: constants[key]})
-            if len(referenced_keys) == last_count:
-                break
-
-        # keep outputs constants
-        outputs_keys = [key for key, value in list(constants.items()) if value["source_type"] == "component_outputs"]
-        referenced_keys = list(set(referenced_keys + outputs_keys))
-        pipeline_tree[PE.outputs] = [key for key in pipeline_tree[PE.outputs] if key in referenced_keys]
-
-        # rebuild constants index
-        referenced_keys.sort(key=lambda x: constants[x]["index"])
-        new_constants = {}
-        for index, key in enumerate(referenced_keys):
-            value = constants[key]
-            value["index"] = index
-            # delete constant reference info to task node
-            for act_id in exclude_task_nodes_id:
-                if act_id in value["source_info"]:
-                    value["source_info"].pop(act_id)
-            new_constants[key] = value
-        pipeline_tree[PE.constants] = new_constants
-
-    @staticmethod
-    def _try_to_ignore_parallel(parallel, converge_id, lines, locations, pipeline_tree):
-
-        ignore_whole_parallel = True
-        converge = pipeline_tree[PE.gateways][converge_id]
-        parallel_outgoing = deepcopy(parallel[PE.outgoing])
-
-        for outgoing_id in parallel_outgoing:
-            # meet not converge node
-            if pipeline_tree[PE.flows][outgoing_id][PE.target] != converge_id:
-                ignore_whole_parallel = False
-                continue
-
-            # remove boring sequence
-            converge[PE.incoming].remove(outgoing_id)
-            parallel[PE.outgoing].remove(outgoing_id)
-            pipeline_tree[PE.flows].pop(outgoing_id)
-            lines.pop(outgoing_id)
-
-        if not ignore_whole_parallel:
-            return
-
-        target_of_converge = pipeline_tree[PE.flows][converge[PE.outgoing]][PE.target]
-        next_node_of_converge = (
-            pipeline_tree[PE.activities].get(target_of_converge)
-            or pipeline_tree[PE.gateways].get(target_of_converge)
-            or pipeline_tree[PE.end_event]
-        )
-
-        # remove converge outgoing
-        lines.pop(converge[PE.outgoing])
-        pipeline_tree[PE.flows].pop(converge[PE.outgoing])
-
-        # sequences not come from parallel to be removed
-        new_incoming_list = []
-        # redirect converge rerun incoming
-        for incoming in converge[PE.incoming]:
-            pipeline_tree[PE.flows][incoming][PE.target] = target_of_converge
-            lines[incoming][PE.target]["id"] = target_of_converge
-            new_incoming_list.append(incoming)
-
-        # redirect parallel rerun incoming
-        gateway_incoming = parallel[PE.incoming]
-        gateway_incoming = gateway_incoming if isinstance(gateway_incoming, list) else [gateway_incoming]
-        for incoming in gateway_incoming:
-            pipeline_tree[PE.flows][incoming][PE.target] = target_of_converge
-            lines[incoming][PE.target]["id"] = target_of_converge
-            new_incoming_list.append(incoming)
-
-        # process next node's incoming
-        TaskFlowInstanceManager._replace_node_incoming(
-            next_node=next_node_of_converge, replaced_incoming=converge[PE.outgoing], new_incoming=new_incoming_list
-        )
-
-        # remove parallel and converge
-        pipeline_tree[PE.gateways].pop(parallel["id"])
-        pipeline_tree[PE.gateways].pop(converge["id"])
-        locations.pop(parallel["id"])
-        locations.pop(converge["id"])
-
-    @staticmethod
-    def _remove_useless_parallel(pipeline_tree, lines, locations):
-        copy_tree = deepcopy(pipeline_tree)
-
-        for act in list(copy_tree["activities"].values()):
-            format_node_io_to_list(act, o=False)
-
-        for gateway in list(copy_tree["gateways"].values()):
-            format_node_io_to_list(gateway, o=False)
-
-        format_node_io_to_list(copy_tree["end_event"], o=False)
-
-        converges = validate_gateways(copy_tree)
-
-        while True:
-
-            gateway_count = len(pipeline_tree[PE.gateways])
-
-            for converge_id, converged_list in list(converges.items()):
-
-                for converged in converged_list:
-
-                    gateway = pipeline_tree[PE.gateways].get(converged)
-
-                    if not gateway:  # had been removed
-                        continue
-
-                    # conditional parallel gateway do not need to trim
-                    is_parallel = gateway[PE.type] == PE.ParallelGateway
-
-                    # only process parallel gateway
-                    if not is_parallel:
-                        continue
-
-                    TaskFlowInstanceManager._try_to_ignore_parallel(
-                        parallel=gateway,
-                        converge_id=converge_id,
-                        lines=lines,
-                        locations=locations,
-                        pipeline_tree=pipeline_tree,
-                    )
-
-            if gateway_count == len(pipeline_tree[PE.gateways]):
-                break
-
-    @staticmethod
-    def preview_pipeline_tree_exclude_task_nodes(pipeline_tree, exclude_task_nodes_id=None):
-        if exclude_task_nodes_id is None:
-            exclude_task_nodes_id = []
-
-        locations = {item["id"]: item for item in pipeline_tree.get("location", [])}
-        lines = {item["id"]: item for item in pipeline_tree.get("line", [])}
-
-        for act_id in exclude_task_nodes_id:
-            if act_id not in pipeline_tree[PE.activities]:
-                error = "task node[id=%s] is not in template pipeline tree" % act_id
-                raise Exception(error)
-
-            act = pipeline_tree[PE.activities].pop(act_id)
-
-            if not act["optional"]:
-                error = "task node[id=%s] is not optional" % act_id
-                raise Exception(error)
-
-            TaskFlowInstanceManager._ignore_act(act=act, locations=locations, lines=lines, pipeline_tree=pipeline_tree)
-
-        TaskFlowInstanceManager._remove_useless_parallel(pipeline_tree, lines, locations)
-
-        pipeline_tree["line"] = list(lines.values())
-        pipeline_tree["location"] = list(locations.values())
-
-        TaskFlowInstanceManager._remove_useless_constants(
-            exclude_task_nodes_id=exclude_task_nodes_id, pipeline_tree=pipeline_tree
-        )
-
-        return True
 
     def creator_for(self, id):
         qs = self.filter(id=id).values("pipeline_instance__creator")
@@ -1454,62 +1170,6 @@ def get_instance_context(pipeline_instance, data_type, username=""):
             }
         )
         return context
-
-
-def preview_template_tree(project_id, template_source, template_id, version, exclude_task_nodes_id):
-    if template_source == PROJECT:
-        template = TaskTemplate.objects.get(pk=template_id, is_deleted=False, project_id=project_id)
-    else:
-        template = CommonTemplate.objects.get(pk=template_id, is_deleted=False)
-    pipeline_tree = template.get_pipeline_tree_by_version(version)
-    template_constants = deepcopy(pipeline_tree["constants"])
-    TaskFlowInstance.objects.preview_pipeline_tree_exclude_task_nodes(pipeline_tree, exclude_task_nodes_id)
-
-    constants_not_referred = {
-        key: value for key, value in list(template_constants.items()) if key not in pipeline_tree["constants"]
-    }
-
-    return {"pipeline_tree": pipeline_tree, "constants_not_referred": constants_not_referred}
-
-
-def get_template_exclude_task_nodes_with_schemes(template_nodes_set, scheme_id_list):
-    """
-    根据执行方案获取要剔除的模版节点
-    @param template_nodes_set:
-    @param scheme_id_list:
-    @return:
-    """
-    exclude_task_nodes_id = []
-    if scheme_id_list:
-        scheme_dict = TemplateScheme.objects.in_bulk(scheme_id_list)
-        scheme_data_set = set()
-        for scheme in scheme_dict.values():
-            scheme_data = json.loads(scheme.data)
-            scheme_data_set.update(scheme_data)
-        exclude_task_nodes_id = list(template_nodes_set - scheme_data_set)
-
-    return exclude_task_nodes_id
-
-
-def preview_template_tree_with_schemes(project_id, template_source, template_id, version, scheme_id_list):
-    if template_source == PROJECT:
-        template = TaskTemplate.objects.get(pk=template_id, is_deleted=False, project_id=project_id)
-    else:
-        template = CommonTemplate.objects.get(pk=template_id, is_deleted=False)
-
-    pipeline_tree = template.get_pipeline_tree_by_version(version)
-    template_constants = deepcopy(pipeline_tree["constants"])
-    template_nodes_set = set(pipeline_tree["activities"].keys())
-
-    exclude_task_nodes_id = get_template_exclude_task_nodes_with_schemes(template_nodes_set, scheme_id_list)
-
-    TaskFlowInstance.objects.preview_pipeline_tree_exclude_task_nodes(pipeline_tree, exclude_task_nodes_id)
-
-    constants_not_referred = {
-        key: value for key, value in list(template_constants.items()) if key not in pipeline_tree["constants"]
-    }
-
-    return {"pipeline_tree": pipeline_tree, "constants_not_referred": constants_not_referred}
 
 
 class TaskOperationTimesConfig(models.Model):
